Many files!
Limitations: All file names have to have this-kind-of-format.md, and they have to be in the 'src/content' folder along with home.md, which is rendered as the default root view.

This is just for the initial prototype; We want this all on the server later. I just want to be able to demo moving between files.